backend/requirements.txt
# Core Dependencies
google-genai>=0.2.0
arxiv>=2.1.0
pymupdf>=1.23.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.0

# Jupyter Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
nbformat>=5.8.0

# Streamlit Dashboard
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# PDF Export (optional for dashboard)
fpdf2>=2.7.0

# Additional utilities
scikit-learn>=1.3.0

#FastAPI
fastapi[standard]